+++ src/datatypes.hpp
@@ -1,7 +1,7 @@
 /*
    datatypes.hpp : Datatype declarations
  
-   Copyright (c) 2018 Simon D. Levy
+   Copyright (c) 2019 Simon D. Levy, Juan Gallostra Acin, Pep Marti-Saumell
 
    This file is part of Hackflight.