+++ src/filters/eskf_struct.hpp
@@ -5,7 +5,8 @@
     You should #include this file after using #define for NNsta (nominal states),
     NEsta (error states) and Mobs (observations)
 
-    Copyright (c) 2018 Simon D. Levy, Juan Gallostra Acin, Pep Martí Saumell
+    Copyright (c) 2019 BonaDrone (www.bonadrone.com)
+    Developed by: Pep Marti-Saumell (jmarti<at>bonadrone.com>) & Juan Gallostra Acin (jgallostra<at>bonadrone.com)
 
     This file is part of Hackflight.
 
@@ -24,8 +25,8 @@
 
 #pragma once
 
-#define NNsta 7
-#define NEsta 6
+#define NNsta 9
+#define NEsta 8
 #define Mobs 3
 
 namespace hf {