+++ src/main.hpp
@@ -4,7 +4,8 @@
    runs it. It acts as a wrapper that automatically creates the correct Hackflight
    instance as a function of the stored parameters.  
 
-   Copyright (c) 2018 Juan Gallostra & Simon D. Levy
+   Copyright (c) 2019 BonaDrone (www.bonadrone.com)
+   Developed by: Pep Marti-Saumell (jmarti<at>bonadrone.com>) & Juan Gallostra Acin (jgallostra<at>bonadrone.com)
 
    This file is part of Hackflight.
 
@@ -152,16 +153,17 @@ namespace hf {
                     h.init(new hf::BonadroneMultiShot(), &rc, &mixer, ratePid);
                 }
                 // Add additional sensors
-                //if (_hasPositioningBoard)
-                //{
-                    //hf::VL53L1X_Rangefinder rangefinder;
-                    //rangefinder.begin();
-                    //h.addSensor(&rangefinder);
+                if (_hasPositioningBoard)
+                {
+                    hf::VL53L1X_Rangefinder rangefinder;
+                    rangefinder.begin();
+                    h.addSensor(&rangefinder);
+                    h.eskf.addSensorESKF(&rangefinder);
                     
                     //hf::OpticalFlow opticalflow;
                     //opticalflow.begin();
                     //h.addSensor(&opticalflow);                    
-                //}
+                }
                 
             } // init