+++ CHANGES.md
@@ -6,8 +6,13 @@
 
 * #37 Hackflight Parametrization
 * #42 ESKF estimates gyrometer bias
+* #43 ESKF estimates vertical position and velocity
+* #43 Add IMU to the list of available sensors 
+* #43 Allow to define the frequency at which sensors correct/update ESKF estimates
 
 ### Changed
+
+* #43 Update file headers
 
 ### Removed
 
@@ -15,4 +20,3 @@
 
 * #39 Firmware version
 * #41 Serial communications with the platform
-